feat: add support for setting `setForceDarkAllowed` on Android via the `accessibilityIgnoresInvertColors` prop
On iOS, there's a recently added prop for views called `accessibilityIgnoresInvertColors`. This is used with nearly the exact inverse logic of Android's new `setForceDarkAllowed` function. This commit makes Android respond to this prop now with the inverse. So for instance if you want to allow forceDark, set it to true. If you want to then disallow it for a child, set it to false.